Add inital Galileo port
This is mostly a stub port for Intel's Galileo board. This commit just places
everything in place so we can start filling the gaps. A few files skeleton were copied
from the native port, but were heavily cleaned up.

For now, the Galileo port will use cpu/x86/ since we are not sure there will be
need for a cpu/quark/ port.

You can test it on Linux with:
$  cd examples/hello-world/ && make TARGET=galileo && ./hello-world.galileo

This will run as a native GNU/Linux application, linked against glibc.