[galileo] Link against Newlib
This commit makes the galileo port compile and link against newlib instead of
glibc. To achieve that we tell both the compiler and the linker to stop using
the system includes and libraries.

The interface with newlib are all the stub syscalls added to newlib-syscalls.c.

The Makefile.x86 was modified for two reasons:
1- the elfloader-x86.c is actually a POSIX only implementation due to the include
of the <sys/mman.h> header;
2- we may not need the elfloader at all.

If both of the above points are confirmed, then a clean solution would be to move
the addition of that source file to platform/native/Makefile.native instead.