[galileo] Add build_newlib.sh and multiboot grub.efi
This commit adds a prebuilt binary of grub2 efi as a "board support package" and
a build_newlib.sh script to download, prepare, build and setup newlib with the
flags and targets we need.

This commit can be reverted when we decide how the galileo port will provide its toolchain.

The newlib version we currently test is 2.2.0-1.