[galileo] Add Multiboot header and custom linker script
This commit adds a loader.S assembly file which is now
the entry point of our kernel. This file adds the multiboot
header to a .multiboot section and calls C function main from
its start label.

This label, in turn, is used as the entry point on the galileo.ld
linker script.