galileo: Use raw binary format
This patch changes Galileo's primary image from ELF (stripped) to raw
binary format since the latter turned out to be more space-efficient.
In order to make the raw binary image bootable, it was required some
changes on the multiboot header and linker script.

In the linker script, we define a few symbols in specific places to
indicate the multiboot header, the begining of text section, end of data
section, and end of bss section.

In multiboot header, we now set the AOUT_KLUDGE flag and address fields.
This flag indicates to the bootloader it should use the address fields
ithin the header to be able to properly load the OS.

After this patch, as shown bellow, Contiki image size reduces ~4kB.
33932 hello-world.elf.striped
29240 hello-world.galileo

Finally, this patch also changes Makefile.galilo to keep printing the
size of the stripped ELF image. This way, we can easily compare the size
of both images after each build.