galileo: Simplify output binaries generation
This commit simplifies the binaries generated when building to the galileo
platform by keeping only a raw binary, a stripped elf and a debugging symbols
file.

It also fixes the binaries naming so the target's platform name is
always part of the output binaries, which is aligned with all other Contiki ports.