galileo: Initialize System Clock using the PIT timer
This commit adds an 8254 PIT driver and implements clock_init(),
initializing the system clock.

The timer used is the legacy PIT chip since its maximum clock resolution
is more than enough to us for now. Moreover, this is the simplest periodic
timer available, with a very straightforward setup. The HPET timer,
in contrast, demands that you first detect its base address by probing the ACPI.

The counter 0 from PIT is used since it is the attached to an interrupt (IRQ0).
It is configured as a square wave generator (mode 3), as this is the most
suitable wave form for periodic clock signals. A 200Hz tick rate is configured
on contiki-conf.h (CLOCK_CONF_SECOND), which means that a timer interrupt
will be generated every 5ms. For comparison, other ports use ~7ms.

This patch also updates the remaining stubs from clock.c, adding TODOs,
removing the deprecated clock_delay() function and adding clock_delay_usec().

More information:
- Quark x1000 Datasheet, section 21.8 (page 873).
- http://wiki.osdev.org/Programmable_Interval_Timer