galileo: Implement mainloop
This patch implements the mainloop for galileo platform. For now, the
mainloop is very simple. All we have to do is call process_run()
repeatedly.

In future, we'll probably improve this mainloop by adding features such
as entering in power save mode when all events have been processed and
kicking a watchdog to be able to recover from abnormal faults.

Additionally, as all hardware peripherals are expected to be initialized
before starting process, this patch also enables IRQs just before
autostart_start().