x86: Add Real Time Clock Driver
This patch adds a driver for Real Time Clock (RTC). The RTC timer is
suitable to implement some operating system features such as the
system clock. Actually, the RTC will be used to implement the system
clock in galileo platform.

The driver introduced by this patch programs the RTC to generate
interrupt periodically. The interrupt frequency can be configured by the
user. The RTC has very specific values for frequency: 2, 4, 8, 16, 32,
64, 128, 256, 512, 1024, 2048, 4096, and 8192 Hz. If the user selects a
different frequency value, timing won't work properly.

On each RTC interrupt, a callback configured by the user is called. As
expected, that callback is executed in interrupt context so the user
should be aware of what it is not supposed to do (e.g. to call blocking
functions).