galileo: Use RTC to implement system clock
This patch changes galileo's system clock code to use RTC instead of
PIT. Since PIT has a much higher resolution, we plan to use it to
implement the rtimer library.

This patch also changes the value of CLOCK_CONF_SECOND to 128. This
value seems to be good enough since it is the most used value in
others platforms supported by Contiki. Besides that, the old value
(200) does not work for RTC.