galileo: Implement PIT timer ISR and System Clock APIs
The timer ISR will be called every 1/CLOCK_CONF_SECOND seconds, which
is our system's tick rate. This function, the pit_int_handler(), sends
calls the pit_int_callback() provided by the driver's user and then send
an end-of-interrupt (EOI) to the PIC. Its current implementation, in turn,
just increments the tick counter and calls etimer_request_poll() if
there are any pending timers.

With the current configuration, with an unsigned 32bits tick_count and a
200 Hz tick rate, our counter will wrap around after roughly 248 days.

This patch also implements clock_time(), which returns the number of
ticks since boot, clock_seconds(), which returns the number of seconds
since boot and clock_wait(), which has to wait for a given number of
ticks t. For the latter, we loop until the current tick_count was increased
by t since the function call.

Issues marked as FIXME are all related to missing APIs on the PIC driver
so they will be addressed by a future commit.