x86: Modify SET_INTERRUPT_HANDLER to enable static handlers to be used
This commit modifies the macro above from interrupt.h, so pointers to static
function can be passed to it as the "handler" argument without the usage of
__attribute__((__used__)) on the function's declaration.

In order to achieve that, the macro now passes the function pointer to the
inline assembly as an argument instead of using its string value directly.
The %P0 is necessary so gcc doesn't generate an indirect call on the output
assembly.