x86: Disable NMI while initializing RTC
According to [1], we should disable non-maskable and maskable interrupts
while initializing RTC. Otherwise, the RTC may be left in an undefined
state (non-functional) if an interrupt occurs. Currently, maskable
interrupts are already disabled, but NMI is not. So this patch changes
rtc_init() to disable NMI.

Since NMI and RTC share the same IO port, just calling nmi_disable()
and nmi_enable is not enough. We should also set the NMI_ENABLE bit
while selecting the RTC_INDEX.

Finally, the nmi_disable() call is not strictly required since we set
the NMI_ENABLE bit while selecting the RTC_INDEX. However, to make clear
that we are disabling NMI and to improve readability (by matching NMI
disable/enable), the nmi_disable() call was purposely used.

[1] http://wiki.osdev.org/RTC