x86: Improve SET_INTERRUPT_HANDLER macro
This patch rewrites the SET_INTERRUPT_HANDLER macro to address the
following issue: Once the macro expands, it will create a global symbol
called  "isr<num>" where <num> is a parameter from the macro. This is not
a problem when "num" is a hard-coded number. However, if the parameter
"num" is a variable or another macro, we might have a problem. For
instance, consider the SET_INTERRUPT_HANDLER in lapic_init_timer(). It
will create the symbol "isrvector". If another function uses this macro
with "vector" as "num" parameter (what is quite likely), we will get a
multiple definition error during linking time.

At the end of the day, we don't care about the symbol's name but it has
to be unique in the whole program. So, to achieve that, we use the token
"%=" from gcc extended asm. This token outputs a number that is unique
to each instance of the asm statement in the entire compilation.

Since the resulting symbol "isr<num>" could be very misleading, the macro
now creates the symbol "trampoline<num>" instead and a XXX comment was
added to warn others programmers that "<num>" is NOT related to the
interrupt number at all. Also, we remove the global scope of this symbol
since it is used locally only.

Because the way the SET_INTERRUPT_HANDLER macro was rewritten, the nested
function _dummy<num>() doesn't make sense anymore and it was removed. We
use a jump instruction to jump over the trampoline code while
registering the interrupt handler.

Finally, this patch also removes trailing backslashs from all macros
in interrupt.h.