x86: Extend PCI driver module to support Ethernet driver
This patch adds the 'meta' field to the generic driver structure to
point to optional driver-defined metadata.  It also modifies the
associated initialization routine to populate it and updates the 16X50
UART driver to use the new initialization routine signature.  This
patch also adds definitions for the PCI Command configuration register
address and a function to perform PCI configuration register writes.