galileo: Omit exception handling unwind tables to save space
This patch modifies the newlib and Contiki C and C++ compiler flags to
omit exception handling unwind tables (see
http://wiki.dwarfstd.org/index.php?title=Exception_Handling).
Removing these tables saves space and has not caused any
readily-apparent functional changes.

Here is the size listing for an example program built without this
patch:
   text	   data	    bss	    dec	    hex	filename
  19825	   1160	  12952	  33937	   8491	all-timers.galileo

Here is the size listing for the same program with with this patch:
   text	   data	    bss	    dec	    hex	filename
  12617	   1160	  12952	  26729	   6869	all-timers.galileo

This patch has a secondary motivation when UEFI support is enabled, as
provided by other patches.  The .eh_frame and .eh_frame_hdr sections
that are otherwise generated are treated as code sections by the UEFI
GenFw program, since they are read-only alloc sections.  They get
grouped with the actual code sections, ahead of the data sections.
This perturbs symbols and complicates debugging.